fix RepositoryLanguage score type

+++ src/models/repository_language.cr
@@ -6,5 +6,5 @@ class RepositoryLanguage
   belongs_to repository : Repository
   belongs_to language : Language
 
-  column score : Float32
+  column score : Float64
 end
